Move RowSortExpressionContext to top level
