+++ presto-main/src/main/java/com/facebook/presto/operator/PagesIndex.java
@@ -24,7 +24,7 @@ import com.facebook.presto.sql.gen.JoinCompiler;
 import com.facebook.presto.sql.gen.JoinCompiler.LookupSourceSupplierFactory;
 import com.facebook.presto.sql.gen.JoinFilterFunctionCompiler.JoinFilterFunctionFactory;
 import com.facebook.presto.sql.gen.OrderingCompiler;
-import com.facebook.presto.sql.planner.SortExpressionExtractor;
+import com.facebook.presto.sql.planner.RowSortExpressionContext;
 import com.google.common.collect.ImmutableList;
 import io.airlift.log.Logger;
 import io.airlift.slice.Slice;
@@ -424,7 +424,7 @@ public class PagesIndex
             //        OUTER joins into NestedLoopsJoin and remove "type == INNER" condition in LocalExecutionPlanner.visitJoin()
 
             try {
-                Optional<SortExpressionExtractor.RowSortExpressionContext> sortExpressionContext = Optional.empty();
+                Optional<RowSortExpressionContext> sortExpressionContext = Optional.empty();
                 if (filterFunctionFactory.isPresent()) {
                     sortExpressionContext = filterFunctionFactory.get().getSortExpressionContext();
                 }
